Arm backend: move evaluation logic out of aot_arm_compiler
A start to simplify the aot_arm_compiler script.
